[Python test] Check that sidechain balance is updated by mempool
Added a Python test to check that whenever the mempool accepts a transaction containing any CSW input, the balance of the sidechain referred by such input is updated accordingly.

For instance, supposing that there is a sidechain with a balance of 10 coins, if the mempool contains 10 CSW inputs withdrawing 1 coin each, the resulting sidechain amount should be considered to be 0 coins. For this reason, any additional incoming CSW input must be rejected for that sidechain.